refactor: extract core simple contracts
